# perf: use `/bin/sh` to launch popup sessions