--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -34,9 +34,10 @@ are noticeable to end-users since the last release. For developers, this project
 
 ### Changed
 
-- Reduce the lantency of popup toggles by up to 50% ([#33])
+- Reduce the lantency of popup toggles by up to 60% ([#33], [#34])
 
 [#33]: https://github.com/loichyan/tmux-toggle-popup/pull/33
+[#34]: https://github.com/loichyan/tmux-toggle-popup/pull/34
 
 ## [0.4.1] - 2025-05-07