--- a/src/eval.sh
+++ b/src/eval.sh
@@ -1,4 +1,0 @@
-#!/bin/sh
-script=${__eval:-}
-unset -v __eval
-eval "$script"