Temporarily change `default-shell` to `/bin/sh` to execute the shell script that launches popup sessions:

1. It addresses the issue described in #14.
2. It should not bring the regressions reported in #15.
3. It also provides a performance boost when opening a popup, since the user shell may take some time to customize before it's ready to run commands.

On my laptop, this PR reduces the latency by ~40% when opening popups.

Before:

```console
$ hyperfine --shell=/bin/sh --warmup 3 "./src/toggle.sh -E --toggle-mode=force-open --on-init='detach'"
Benchmark 1: ./src/toggle.sh -E --toggle-mode=force-open --on-init='detach'
  Time (mean ± σ):      48.7 ms ±   2.0 ms    [User: 7.2 ms, System: 7.7 ms]
  Range (min … max):    44.4 ms …  55.0 ms    61 runs
```

After:

```console
$ hyperfine --shell=/bin/sh --warmup 3 "./src/toggle.sh -E --toggle-mode=force-open --on-init='detach'"
Benchmark 1: ./src/toggle.sh -E --toggle-mode=force-open --on-init='detach'
  Time (mean ± σ):      27.7 ms ±   4.7 ms    [User: 7.6 ms, System: 8.6 ms]
  Range (min … max):    24.2 ms …  43.9 ms    66 runs
```